refactor: move `getDashboardPremiumPerks` to `validator_dashboard.go`
- since it is a validator dashboard helper and using the existing `withFailing` helps testing
